% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkwhale.R
\name{checkWhale}
\alias{checkWhale}
\title{Check if a particular whale type has been sighted}
\usage{
checkWhale(x)
}
\arguments{
\item{x}{A whale type (A character string in quotation mark)}
}
\value{
A character string telling whether or not the input whale
type has occurred in recent sightings.
If yes, it will return the specfic number of sightings.
}
\description{
By passing a particular whale type to the function, it will tell you if
this type of whale has occurred in the recent 1000 sightings
(maximum of sightings to access). If yes, it will return the
specfic number of sightings.If not, it will tell you
there are no such records.
}
\examples{
checkWhale("orca")
checkWhale("humpback")
}
