% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whale_dist.R
\name{whale_dist}
\alias{whale_dist}
\title{Get whale sightings that occured near a center point within a radius}
\usage{
whale_dist(x, y, z)
}
\arguments{
\item{x}{Latitude (Must be within the range [-90, 90])}

\item{y}{Longitude (Must be within the range [-180, 180])}

\item{z}{Radius (Must be a non-negative number)}
}
\value{
A dataframe that contains the sightings falling in the specified region
}
\description{
Users can specify the location of sightings they want to get. There are three parameters they can modify:
latitude, longitude, and radius. A pair of latitude and longitude fixes a center point and the radius refers
to the distance around that center point. Togther, they forms a region. By passing latitude, longitude,
and radius sequentially to the function, it will return a R dataframe that contains sightings falling in
specified region, as well as a variety of information about the sightings including whale species,
sighting latitude & longitude, location, and sighting date & time.
}
\examples{
whale_dist(48.5159, -123.1524, 0.5)
whale_dist(37.765, 120.98, 120)
}
