% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whale_info.R
\name{whale_info}
\alias{whale_info}
\title{Get comprehensive information about the most recent whale sightings}
\usage{
whale_info(x)
}
\arguments{
\item{x}{An integer between 1 and 1000 that indicates the number of recent sightings you want to get}
}
\value{
A R dataframe that contains a variety of information about the whale sightings, including
whale species, sighting latitude & longitude, location, and sighting date & time.
}
\description{
The Whale Hotline API allows users to get a maximum of 1000 recent whale sighting records.
By pass an integer x between 1 and 1000 to the function, it will return a R dataframe that contains
x recent whale sightings with detailed information including whale species, sighting latitude & longitude,
location, and sighting date & time.
}
\examples{
whale_info(500)
whale_info(1000)
}
